feat: create reusable modal, create progress components
